requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
arango==0.2.1
asttokens==3.0.0
attrs==23.2.0
autograd==1.7.0
autograd-gamma==0.5.0
ax-platform==0.5.0
beautifulsoup4==4.13.3
biomart==0.9.2
biopython==1.85
botorch==0.13.0
Brotli==1.1.0
build==1.2.2.post1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.1
DeepPurpose==0.1.5
descriptastorus @ git+https://github.com/bp-kelley/descriptastorus@9a190343bcd3cfd35142d378d952613bcac40797
distro==1.9.0
dm-tree==0.1.8
editdistance==0.8.1
einops==0.8.1
executing==2.2.0
-e git+https://github.com/SigmaGenX/TamGen@9459b3f0290ea7b79061d9e81eae284f93891a82#egg=fairseq
faiss-cpu==1.10.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.14.0
fonttools==4.56.0
formulaic==1.1.1
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
fy-common-ext==0.1.0
gdown==5.2.0
gmpy2==2.1.5
gpytorch==1.14
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.4
importlib_metadata==8.6.1
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jaxtyping==0.2.38
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.3.3
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.11
lifelines==0.30.0
linear-operator==0.6
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
narwhals==1.29.0
nest-asyncio==1.6.0
networkx==3.4.2
nose==1.3.7
numpy==1.26.4
openai==1.65.2
opt_einsum==3.4.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandas-flavor==0.6.0
parso==0.8.4
pexpect==4.9.0
phenolrs==0.5.9
pillow==10.2.0
pip-tools==7.4.1
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
prettytable==3.15.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-arango==8.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
rdkit==2024.3.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.9.9
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
setuptools==72.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
subword-nmt==0.3.8
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
tzdata==2025.1
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
streamlit==1.40.2
langchain-google-genai==2.0.7
pyvis==0.3.2
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-geometric==2.5.3
